[Feat, Design] 랜딩 페이지 신규 생성, 네브바 및 체험 로그인 버튼에 로고 추가, 홈페이지 변동 차트 xLabel 갱신
